pyproject.toml
[project]
name = "api-mckerlie-com"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Adam McKerlie", email = "adammckerlie@gmail.com" }
]
dependencies = [
    "django>=4.2.6",
    "langchain>=0.0.325",
    "requests>=2.31.0",
    "openai>=0.28.1",
    "transformers>=4.34.1",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.1",
    "djangorestframework>=3.14.0",
    "unstructured>=0.10.28",
    "markdown>=3.5.1",
    "chromadb>=0.4.15",
    "pypdf>=3.17.0",
    "pdf2image>=1.16.3",
    "pdfminer.six>=20221105",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
